4-letter bird codes in BigQuery
This change makes the csv-to-avro pipline read birds.tsv instead of
birds.txt, with the difference being that the .tsv has unquoted fields.
This simplifies the parsing code. At the same time, it changes which
column of the .tsv is passed through to logit.key so that it is the
4-letter bird code rather than the 6-letter one.

For aru2point, it also makes the parsing silently skip any rows where
point is non numeric. Avro records resulting from such filenames are
still emitted with a null point submessage.