y / yARU renaming
This change is functional, not only cosmetic. Prior to this, the JAGS
model was referencing names not in its (data <- list), yet still it was
running without error.

Since JAGS says y.ind ~ dbern, it is also important that we feed it
binary values. It could be helpful to find a functional form of the
model that makes use of the actual count not just for the ARU samples
but also for the point counts.