Removes thresholding from score samples
A model assumption was violated, and this fixes it. If the scores are
filtered by being above a hard threshold, it is not reasonable to assume
they are distributed as a mixture of 2 Gaussians.

On the other hand, there are too many scores for the model to fit in a
reasonable time if all of them are used. Sampling instead of
thresholding can reduce the number of scores without changing their
distribution.