Manages packages with renv
https://rstudio.github.io/renv/index.html

This commit adds the files generated by renv::init() from a clean
version of our main branch on the date of the commit.

This includes a project-level .Rprofile that "activates" the renv.
Activation installs the renv package for contributors who do not already
have it, after which it installs the packages from renv.lock.

When adding code that relies on a new package, please call
renv::snapshot(), check the updates it makes to renv.lock, and include
them in your commit.