Passes species and years from indices to readMl
This is a bug that did not show itself when this was being called from a
test script where species (plural) had been defined locally but revealed
itself when combined.R was ported.

It is indended that outerIndices specify which species and years to
include and that readMl get to know those in case it can use them to
make the query more efficient.