Generates ‘wide’ versions of RASTER variables from previous commit including new;
RAVG values (from the dnbrCBI4 layer that Becky said to use)

as well as new variables with respect to ‘large trees’:

“Perc_LargeTreeHeight_” == The percent of the plot (1ha or 4ha) that had pixels with trees >= 28m

“Perc_LargeTreeCover_” == The actual coverage **only in the pixels > 28m**

If you multiply the former x the latter (e.g. Perc_LargeTreeHeight_ * Perc_LargeTreeCover_) you get the first -order approximation of the fraction of the entire sample that is covered with large trees, which is a tiny number.