Ability to poll multi-species result frame
The JAGS model takes much, much longer to converge for a few species
than for most of the others. (That is, fit for each species has
stragglers.)

This change makes to possible to getCurrentResult() and see estimates
for those that have finished so far while fitting for the others is
still in progress.